randevents.py file moved to tests/ directory
